logs are now printed during training
Previously, couldn't tell if the network was training at all because nothing was
printed. The initial suspicion was that the network wasn't training and it was
taking an extremely long time to do an inference. Doesn't make sense so I dug
deeper.

Found out that the problem lies with not adding:
tf.logging.set_verbosity(tf.logging.INFO)

Refer to https://www.tensorflow.org/get_started/monitors for more information.