fixed global_step and control_flow_operations
global_step wasn't changing previously and this was fixed by renaming it to
global_step_tensor. I suspect there might be some conflict with the implicit
global_step in estimator.

Following the above change, global_step_tensor started incrementing in steps of
2. This was due to the fact that there is a g_optim and d_optim and even though
I'm only running one at any one time, both cause global_step to increment. This
was resolved by incrementing global_step seperately and grouping it together as
a train_op.

tf.mod, which controls whether or not we optimize the generator or
discriminator, had stochastic behaviour. It was supposed to print repeated
numbers without skipping. However, skipping was observed. It was suspected to be
a tf.control_dependencies problem and this was resolved by executing mod before
the global_step increment.

Unresolved:
1. Generator and Discriminator not training properly.
2. Use the implicit global_step in estimator.