corrected missing reshape in discriminator
Last layer in discriminator involves a dense layer, which wasn't reshaped. It
ended up having 100x4x4x1 logits as output. This was a problem because the
logits labels were simply based on the previous shape. Hence, the neural network
wasn't trying to learn anything sensible.

Unresolved:
1. Generator and Discriminator not training properly. This could potentially
solve it but not tested on GPU yet.
2. Use the implicit global_step in estimator.