added implicit global_step and image summary
tensorflow/python/estimator/estimator.py line 189 has the train function.
Tracing the code leads you to _train_model in line 556 and this leads you to
tensorflow/python/training/training_util.py line 95, where create_global_step is
called. Here, we can see how global_step is created and the various methods that
we can use to get the global_step that is being used by Estimator. The options
are get_global_step or get_or_create_global_step. More information is in the
code itself.

Summaries have been improved such that one simply needs to add tf.summary.image
within the model_fn and the magic is done on the back end. Do note that the
summary steps in RunConfig need to be edited such that you are writing out a
summary at the rate that you want.

Unresolved:
1. Generator and Discriminator not training properly.