edited sampling of zs
Previously, it might be the case that the zs were not sampled properly because I
took a sample and then used tf.train.batch. I do not understand the internal
workings of tf.train.batch but I might be using the same z. Changed it to sample
the entire batch of zs at once. There is a need to verify that all zs are
different.

Unresolved:
1. Generator and Discriminator not training properly.
2. Summaries for relevant operations and z vector.